requirements.txt
streamlit
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
